requirements.txt
asyncio==3.4.3
colorama==0.4.6
deepdiff==8.1.1
dipy==1.10.0
h5py==3.12.1
nibabel==5.3.2
numpy==2.2.1
opencv-python==4.11.0.86
orderly-set==5.2.3
packaging==24.2
pygame==2.6.1
python-dotenv==1.0.1
scipy==1.15.1
setuptools==75.8.0
setuptools-scm==8.1.0
TIME-python==0.0.17
tqdm==4.67.1
trx-python==0.3
typing_extensions==4.12.2
websockets==14.1
